requirements.txt
rhino3dm==7.14.2
git+https://github.com/contextmachine/mmcore.git
setuptools>=65.6.3